backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
python-dotenv==1.0.1
pydantic_settings==2.6.0
asyncpg==0.30.0